+++ FastAnimations/Handlers/BreakingGeodeHandler.cs
@@ -1,5 +1,3 @@
-#nullable disable
-
 using Pathoschild.Stardew.FastAnimations.Framework;
 using StardewValley;
 using StardewValley.Menus;