+++ FastAnimations/Handlers/LoadGameMenuHandler.cs
@@ -1,5 +1,3 @@
-#nullable disable
-
 using Pathoschild.Stardew.FastAnimations.Framework;
 using StardewModdingAPI;
 using StardewValley;