+++ FastAnimations/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Added support for horse flute animation.
+* Improved translations. Thanks to ChulkyBow (updated Ukrainian)!
 
 ## 1.9.12
 Released 27 February 2022 for SMAPI 3.13.0 or later.